Lower case event name in _Widgetbase.on
The emit method of _WidgetBase converts the event type (e.g. 'click'), but the on method does not. This leaves us with really annoying behaviour where you can have the following:

//Attach to selectItem
a.on("selectItem", function(){ alert("This should fire"); });

//emit selectItem
a.emit("selectItem", {});

//Nothing happens!! because emit actually emitted 'selectitem'

This change is very simple, it does change behaviour though, and maybe it should only come in with dojo 2.0, but the current behavious is incredibly confusing. The alternative is to stop emit from lowercasing the event type, but having 'selectitem' and 'selectItem' hit different functions is a recipe for disaster.